test(test): complete log upload
The log upload works by opening a file for the current date and
uploading each line in the file. Further improvements may be made by
reading the files recursively and storing the data in a buffer. This
allows the file to be closed and for the data to be uploaded separately.