test(main): Working demo
The device will record data as the user moves, and record interrupt
signals when a Z High low event occurs. The setup will ensure that the
interrupt is unlatched and ready to be triggered. The main loop will
wait until the GPS has connected to at least 3 satellites before data is
collected. Upon data being prepared for upload, a WiFi connection must
be established. It must be noted that the WiFi connection will not time
out and may be caught in an infinite loop.