src(impact) add impact.toLogString()
Impact objects store data related to a vehicle responding to an impact.
This data must be logged and will contain a datetime and location. The
data stored in this object may be written to a file using the
toLogString() method.
The toLogString() method returns a string of the form:
YYYYMMDD,HHMMSS,LATITUDE,LONGITUDE,1