src(dispatcher): initial working observer pattern
The dispatcher will act as both a mediator and an observer. It will
communicate data between interacting objects while listening to state
changes of those objects. When a state change occurs, the dispatcher is
updated of what change occured (state) and in what object. It will then
dispatch a task relevant to that task, communicating data between
interacting objects as needed.