src(data_collector): initialze class and enums
The DataCollector class is responsible for aggregating external
sensors and collecting data from them. The interface provided by the
DataCollector allows the caller to read a specific data type from a
selected sensor using the sensor_t and {sensor_data_t} enums.

The DataCollector will cache data that may be re-read to reduce
computational overhead and improve speed. High volume data may be writte
to a buffer and be batch-collected upon calling the interace method
getData().