src(storage): initialize class
The Storage class is responsible for managing local storage, and
reading and writing files. Data must be written or read to a buffer of
a specifed length. Upon constructing the Storage class, the SD card will
be initialized and will be referenced by the file-system in a private
context, reducing the interface function's complexity.