architecture refactoring
Due to the difficulty of managing the inital implementaion, the
architecutre of this design is being refactored into a managable state
that is easy to read and understand. For this reason, many of the
existing data structures have been refactored into a single structure,
showing all the essential information in a single location. Use cases
have been implemented in their corresponding controllers, and sensor
interfaces have beeen developed to provide a layer of abstraction
between drivers and implementation methods.