src(devices): add device function headers
The devices directory lists all of the devices used in this project. The
accelerometer, environmental_sensor, and gps are sensors responsible for
collecting data and aiding in state identification, while storage and
wifi are used for temporary storage and data transmission.

Controllers will depend upon the information and functions of these
devices to achieve their respective use cases.