src(*),test(main): add main loop structure
The main has been designed around the state of the vehicle.
Vehicle states include the following, and are specified in
gps_vehicle_state_enum.h:
* idle
* moving
* unknown - the vehicle movement state cannot be known if there is no
            gps connection

To improve read speeds, cache has been implemented in the GPS and
environmental sensor.

Cache has been implemented in the GPS for the following:
* year
* month
* day
* hour
* minute

Cache has been implemented in the Environment Sensor for the following:
* temperature
* humidity
* pressure
* altitude