test(acc_rolling_buffer): add dual core test
The ESP32 supports dual core processing. This is necessary for this
implementation as it allows a larger "rolling buffer". This allows us to
constantly poll the accelerometer while performing control operations on
a seperate core.

When an interrupt occurs, the device will collect the rolling data
buffer, informing us of pre-impact data. The device will proceed to
collect post-impact data until this post_data buffer is full.

Using globals and latched interrupts, the main control core will know an
interrupt occured and collect this impact data and log it to a file. The
control core will then unlatch the interrupt, allowing this process to
be repeated.