src(accelerometer): replace driver
The existing LIS3DH driver did not allow for rapid reads above 1 KHz. If
this was attempted, the accelerometer woudl return 0 values for readings
and had to be reset to resume returning correct values.

For this reason, the driver has been replaced with Adafruit's
Adafruit_LIS3DH accelerometer library.

Testing has verified this driver works for reading at high frequencies.