src(test/*): test new accelerometer driver
The new Adafruit_LIS3DH accelerometer has been tested by reading at high
frequencies.

Data is captured into three buffers of length 1600. These buffers each
store data from a single accelerometer axis (x, y, or z).

This data has also been tested with storage, allowing for all 3 buffers
to be written in CSV format to a file on the SD card (see
test/storage/storage.ino). The resulting file is approximately 30KB in
size.

A rolling buffer was tested, however the implementation at this time
does not achieve the desired result. A single buffer for each axis
storing the most recent 1600 samples appears to be sufficient. This may
be refactored after testing the deployed device.

Multi-core processing has been tested wit the accelerometer and file
writing to the SD card. However, writing to the SD card results in an
unhandled exception and causes the device to reset. This issue may be
investigated later if multi-core processing is found to be important.

Finally, a file header has been added to the main test program.